requirements.txt
asgiref==3.2.10
Django==3.1
django-crispy-forms==1.9.2
django-chunked-upload==2.0.0
django-filter==2.3.0
django-filter==2.3.0
django-login-required-middleware==0.4
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
django-rest-passwordreset==1.1.0
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
django-cors-headers==3.5.0
django-allauth~=0.42.0
django-rest-auth~=0.9.5
python-dotenv==0.14.0
PyPDF2==1.26.0
reportlab==3.5.57
